+++ package/core/Fetcher.d.ts
@@ -1,4 +1,4 @@
 import { YTDL_RequestOptions } from '../types/Options';
 export default class Fetcher {
-    static request<T = unknown>(url: string, { requestOptions, rewriteRequest }?: YTDL_RequestOptions): Promise<T>;
+    static request<T = unknown>(url: string, { requestOptions, rewriteRequest, originalProxyUrl }?: YTDL_RequestOptions): Promise<T>;
 }