+++ package/core/Fetcher.js
@@ -3,11 +3,15 @@ Object.defineProperty(exports, "__esModule", { value: true });
 const undici_1 = require("undici");
 const errors_1 = require("./errors");
 class Fetcher {
-    static async request(url, { requestOptions, rewriteRequest } = {}) {
+    static async request(url, { requestOptions, rewriteRequest, originalProxyUrl } = {}) {
         if (typeof rewriteRequest === 'function') {
             const WROTE_REQUEST = rewriteRequest(url, requestOptions);
             requestOptions = WROTE_REQUEST.options;
             url = WROTE_REQUEST.url;
+        }
+        if (originalProxyUrl) {
+            const PARSED = new URL(originalProxyUrl);
+            url = `${PARSED.protocol}//${PARSED.host}/?url=${encodeURIComponent(url)}`;
         }
         const REQUEST_RESULTS = await (0, undici_1.request)(url, requestOptions), STATUS_CODE = REQUEST_RESULTS.statusCode.toString(), LOCATION = REQUEST_RESULTS.headers['location'] || null;
         if (STATUS_CODE.startsWith('2')) {
@@ -18,7 +22,7 @@ class Fetcher {
             return REQUEST_RESULTS.body.text();
         }
         else if (STATUS_CODE.startsWith('3') && LOCATION) {
-            return this.request(LOCATION.toString(), { requestOptions, rewriteRequest });
+            return this.request(LOCATION.toString(), { requestOptions, rewriteRequest, originalProxyUrl });
         }
         const ERROR = new errors_1.RequestError(`Status Code: ${STATUS_CODE}`);
         ERROR.statusCode = REQUEST_RESULTS.statusCode;