+++ package/core/clients/Base.js
@@ -37,6 +37,7 @@ class Base {
                     body: typeof requestOptions.payload === 'string' ? requestOptions.payload : JSON.stringify(requestOptions.payload),
                 },
                 rewriteRequest: params.options.rewriteRequest,
+                originalProxyUrl: params.options.originalProxyUrl,
             }, IS_NEXT_API = url.includes('/next');
             try {
                 Fetcher_1.default.request(url, OPTS)